Modify the way of ticker.add to remove input lag
In the previous code, "pikaVolley.gameLoop()" was called at the first moment
of each tick and "renderer.render(stage)" was called at the last moment of each
tick, so, there is always about 40ms (in 25 FPS) gap between the two function calls.
Change the way of ticker.add to remove this gap.